added delete recipe img in storage

+++ src/api/services/recipes.service.ts
@@ -1,7 +1,8 @@
 import MessageResponse from "../../interfaces/MessageResponse";
-import { uploadBase64ImageToStorage } from "../../utils/uploadImage";
+import { base64ToArrayBuffer, hashString } from "../../utils/helperFunctions";
 import * as recipeOperationsDB from "../data-access/recipe.da";
 import { RecipeDocument } from "../models/recipe.model";
+import { firebaseStorageOperations } from "./firebase.service";
 
 export const recipeOperations = {
     getAll: async (userId: string): Promise<RecipeDocument[]> => {
@@ -10,7 +11,13 @@ export const recipeOperations = {
     },
 
     addRecipe: async (recipe: RecipeDocument): Promise<RecipeDocument> => {
-        const image_url = await uploadBase64ImageToStorage(recipe.image_url, Date.now().toString())
+        // Extract the base64 part
+        const base64Image = recipe.image_url.replace(/^data:image\/(png|jpeg);base64,/, '');
+
+        // Convert base64 to ArrayBuffer
+        const imageBuffer = base64ToArrayBuffer(base64Image);
+
+        const image_url = await firebaseStorageOperations.uploadImage(imageBuffer, hashString(recipe.recipe.description))
 
         const newRecipe = await recipeOperationsDB.addRecipe({...recipe, image_url} as RecipeDocument)
 
@@ -18,7 +25,21 @@ export const recipeOperations = {
     },
 
     deleteRecipe: async (userId: string, recipeId: string): Promise<MessageResponse> => {
-        await recipeOperationsDB.deleteRecipe(recipeId)
+        const recipe = await recipeOperationsDB.getRecipe(recipeId)
+
+        if (!recipe) {
+            throw new Error('Recipe not found')
+        }
+
+        if (recipe.userId !== userId) {
+            throw new Error('You are not authorized to delete this recipe')
+        }
+
+        await Promise.all([
+            firebaseStorageOperations.deleteImage(hashString(recipe.recipe.description)),
+            recipeOperationsDB.deleteRecipe(recipeId)
+        ]);
+
         return { message: 'Recipe deleted' }
     },
